[processing] Generalise modeler multi-step feedback proxy for use outside of modeler
This can be useful too for individual algorithms
