+++ python/core/processing/models/qgsprocessingmodelalgorithm.sip
@@ -374,8 +374,6 @@ Translated description of variable
 };
 
 
-
-
 /************************************************************************
  * This file has been generated automatically from                      *
  *                                                                      *
